Build separate shared library for client
By building the shared library in Bazel and linking it from the module,
we can split out and reorganize the client source however TDD guides us.
