+++ module/conf/lauth.conf
@@ -1,6 +1,8 @@
 LoadFile /usr/lib/${cpu_arch}/libssl.so.1.1
 LoadFile /usr/lib/${cpu_arch}/libstdc++.so.6
 LoadModule lauth_module modules/mod_lauth.so
+LoadModule rewrite_module modules/mod_rewrite.so
+LoadModule remoteip_module modules/mod_remoteip.so
 <Location /lauth>
 SetHandler lauth
 </Location>
