+++ module/mod_lauth.cpp
@@ -43,6 +43,7 @@
 #include "ap_config.h"
 
 #include "lauth/client.h"
+
 #include <string>
 
 using namespace mlibrary::lauth::v1;
